updated benchmarks and readme
